backend/requirements.txt
# Django framework
django>=4.2,<5.0

# Django REST Framework for API development
djangorestframework>=3.14,<4.0

# PostgreSQL database adapter for Django
psycopg2-binary>=2.9,<3.0

# CORS support for APIs (if needed for frontend interaction)
django-cors-headers>=4.0,<5.0

# For environment variable management (optional but recommended)
python-decouple>=3.8,<4.0

# For handling JSON Web Tokens (JWT) for authentication (optional)
djangorestframework-simplejwt>=5.2,<6.0

# Linting and code quality (optional)
flake8>=6.1,<7.0

# Testing framework (optional but useful for development)
pytest>=7.4,<8.0
pytest-django>=4.5,<5.0
